backend/pyproject.toml
[project]
name = "youtube-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "fastapi[standard]>=0.100.0",
    "uvicorn>=0.23.0",
    "youtube-transcript-api>=0.6.2",
    "qdrant-client>=1.7.0",
    "notion-client>=2.2.1",
]
